requirements.txt
Django==4.2.4
djangorestframework==3.14.0   
requests==2.31.0
pandas==2.2.3
psycopg2-binary==2.9.9
